clippy::pedantic, with no lint allowed globally — and the two real bugs that policy surfaced (#31)
Enables `clippy::pedantic` workspace-wide with **no lint allowed globally**. Every production warning was resolved on its merits rather than silenced in the manifest.

Two of them turned out to be real bugs. That is what the policy bought.

## The number was 294, and 294 was wrong

The headline count in the design doc is measured on the default build configuration only. Scope D checked the other three and found 14 more:

| surface | warnings |
| --- | --- |
| `tests/` + `examples/` targets (51 files) | 330 |
| inline `#[cfg(test)]` modules in `src/` (54) | 379 |
| **production, default config** | **294** |
| **production, other build configs** | **+14** |

`cargo clippy --workspace` never compiles `redextape-native` under `--no-default-features` or `--features llvm`. 6 of the 14 were real production warnings, 6 were inside `#[cfg(test)]`, 2 were in `--no-default-features` stubs. The spec still leads with 294; the ledger carries the correction.

## Two real bugs, closed as slices with tests

**`redextape-native-rt`'s heap/box indexing** (`db14c89`) — four `(p-1) as usize` casts could alias a valid cell instead of faulting on a 32-bit target. Now `usize::try_from` folded into the *existing* fault path: provably infallible on 64-bit, no new error variant, no `extern "C"` signature or fault-string change.

**`lambda::term::shift`'s upper bound** (`5359833`) — the site had an `assert!` guarding the sign direction and nothing guarding the top. `shift` is `pub`, and `var(u32::MAX)` + `shift(1, 0, ·)` overflows regardless of what type `d` is, so narrowing the signature would not have closed it. Now bounded by a matching assert.

**`viewmodel::LinkIndex::tm_owner`** (`cbe595f`) — `map_or(-1, |n| n as i32)` where `-1` is the live "no owner" sentinel. A `NodeId` past 2^31 casts negative and becomes indistinguishable from "no owner", silently deleting a real link from the UI. Now declines the whole leg via `Option<Vec<i32>>` rather than laundering the overflow into the sentinel. The narrower alternative — a distinct negative sentinel preserving other entries — was weighed and rejected in the field's doc: `web/src/link.ts` collapses *any* negative to `null`, so it renders identically to `-1` at the consumer and tells the UI the same false thing.

Every one of these tests was independently mutated by a reviewer — not merely reverted — to confirm it discriminates a subtly-wrong fix and not just an absent one.

## Policy and the two exemption mechanisms

```toml
pedantic = { level = "warn", priority = -1 }
```

`priority = -1` is not load-bearing today; it is there so a future per-lint `allow` works rather than errors. That matters because `rust-toolchain.toml` tracks unpinned `stable`, and `pedantic` is large and actively extended — the first unexplained red CI after a Rust release is now most likely a new pedantic lint, and this field is the one-line remedy.

Test code is exempt by two different mechanisms, because the two surfaces differ:

- **5 lines** — `#![cfg_attr(test, allow(clippy::pedantic))]` at each crate root covers all 54 inline `#[cfg(test)]` modules. Under `--all-targets` each lib compiles twice and `cfg(test)` holds only in the test-harness pass, so production warnings still surface from the other one. This also reaches the four `#[cfg(all(test, feature = "..."))]` modules that `clippy.toml`'s header notes are invisible to clippy's own in-test detection.
- **51 file-level attributes** — Cargo has no per-target lint config, so `tests/` and `examples/` need one each. 49 already carried an identical `unwrap`/`expect`/`panic` allow to extend; 2 had none.

## A claim this work disproved

`Cargo.toml`'s lint block said the five panic-lints made "no library path may panic" *mechanical*. They do not — `assert!` and `unreachable!` are panics that trip none of them, and no clippy lint bans `assert!` at all. The comment now says what is true, and names `shift` as the known live exception: a `pub` path that panics via two documented asserts, because full compliance would mean threading `Result` through a hot recursive path into `redextape-wasm` and `redextape-native`.

## Verification

```
cargo clippy --workspace --all-targets -- -D warnings   clean, with pedantic on
./scripts/check-all.sh                                  green: base, LLVM, browser
cargo nextest run --workspace                           874 passed, 3 skipped
web: pnpm run build:wasm && pnpm run typecheck          green
```

`check-all.sh`'s LLVM rows need `LLVM_SYS_221_PREFIX=/usr` — LLVM 22.1.8 is installed as the system llvm, not at the `/usr/lib/llvm-22` path the script's probe assumes.

## Known gaps, deliberately left open

- **Nothing bounds Core AST node *count*** (only nesting depth), and **nothing bounds `Program::code.len()`** the way `MAX_TAPES`/`MAX_SLOTS` bound their quantities. Two agents found these independently. The casts that depended on them are safe either way now, so neither is a live defect — but it is one gap wearing two hats and it will surface again.
- **`scripts/check-all.sh` lints feature variants of `redextape-native` only.** It builds and tests `redextape-core --features serde` but never lints it, so serde-gated core code sits outside every lint gate. Benign today — the 3 serde-only warnings are all in an example this branch exempts — but the hole is more consequential now that pedantic is on. A `base|clippy|-p redextape-core --features serde --all-targets` row would close it.

+++ crates/redextape-core/src/tm/machine.rs
@@ -52,6 +52,7 @@ pub struct Machine {
 impl Machine {
     /// The sorted set of concrete symbols appearing in any rule (wildcards excluded). Derived — the
     /// text form and Plan 4's view model present it; it is not stored.
+    #[must_use]
     pub fn alphabet(&self) -> Vec<Symbol> {
         let mut set: BTreeSet<Symbol> = BTreeSet::new();
         for s in &self.states {
@@ -71,10 +72,15 @@ impl Machine {
     /// set `* ; [ ]` and not whitespace (the blank `_` is allowed), every state name is unique (the
     /// text form's identity), and an accept state carries no rules (`print_tm` drops them). Returns
     /// the problems (empty == valid). Never panics.
+    #[must_use]
     pub fn validate(&self) -> Vec<String> {
         let mut errs = Vec::new();
-        let n = self.states.len() as u32;
-        if self.start >= n {
+        // Compare in `usize` space (widen `start`/`next` up, never narrow `states.len()` down):
+        // `states.len() as u32` would truncate past `u32::MAX` states, and a truncated `n` would make
+        // this out-of-range check compare `start`/`next` against the WRONG bound — the same class of
+        // bug `mul_count_unrepresentable` avoids the same way, in `lower_tm.rs`.
+        let n = self.states.len();
+        if self.start as usize >= n {
             errs.push(format!("start state {} out of range (states: {n})", self.start));
         }
         let mut seen_names: HashSet<&str> = HashSet::new();
@@ -95,7 +101,7 @@ impl Machine {
                 if r.read.len() != self.tapes || r.write.len() != self.tapes || r.moves.len() != self.tapes {
                     errs.push(format!("state {i} `{}` rule {j}: arity != {} tapes", s.name, self.tapes));
                 }
-                if r.next >= n {
+                if r.next as usize >= n {
                     errs.push(format!("state {i} `{}` rule {j}: next {} out of range", s.name, r.next));
                 }
                 for sym in r.read.iter().chain(r.write.iter()).flatten() {
